backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
pymupdf
google-genai
python-dotenv
numpy
httpx
